api/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.2.10           # via django
certifi==2020.6.20        # via requests
cffi==1.14.3              # via cryptography
chardet==3.0.4            # via requests
cryptography==3.1.1       # via pyjwt
defusedxml==0.6.0         # via python3-openid
dj-rest-auth==1.1.2       # via -r requirements.in
django-allauth==0.43.0    # via -r requirements.in
django-cors-headers==3.5.0  # via -r requirements.in
django==3.1.2             # via -r requirements.in, dj-rest-auth, django-allauth, django-cors-headers, djangorestframework, djangorestframework-simplejwt
djangorestframework-simplejwt==4.4.0  # via -r requirements.in
djangorestframework==3.12.1  # via -r requirements.in, dj-rest-auth, djangorestframework-simplejwt
idna==2.10                # via requests
oauthlib==3.1.0           # via requests-oauthlib
pycparser==2.20           # via cffi
pyjwt[crypto]==1.7.1      # via django-allauth, djangorestframework-simplejwt
python3-openid==3.2.0     # via django-allauth
pytz==2020.1              # via django
requests-oauthlib==1.3.0  # via django-allauth
requests==2.24.0          # via django-allauth, requests-oauthlib
six==1.15.0               # via cryptography
sqlparse==0.4.1           # via django
urllib3==1.25.11          # via requests
